added nice FE logic + phone number in BE + contacts CRUD operations

+++ frontend/src/app/contact-page/contact-page.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { ChangeDetectorRef, Component, NgZone, OnInit } from '@angular/core';
 import { Contact, ContactService } from '../services/contact-service';
 import { UserInfo, UserService } from '../services/user-service';
 import { AuthService } from '../services/auth-service';
@@ -9,36 +9,40 @@ import { AuthService } from '../services/auth-service';
   templateUrl: './contact-page.html',
   styleUrl: './contact-page.scss',
 })
-export class ContactPage {
+export class ContactPage implements OnInit {
   userInfo: UserInfo | undefined
   userContacts: Contact[] = []
   error = ''
   contactFormEnabled = false;
+  action: 'create' | 'update' = 'update';
   selectedContactId?: number | undefined;
 
   constructor(private userService: UserService, private contactService: ContactService,
-    private authService: AuthService
+    private authService: AuthService, private cdr: ChangeDetectorRef
   ) {}
 
   ngOnInit(): void {
-    
-    this.userService.getUserInfo().subscribe( {
-      next: (info => this.userInfo = info),
-      error: (err: Error) => {
-        this.userInfo = undefined;
-        this.error = err.message;
-      }
-    })
+
+    this.userService.getUserInfo().subscribe({
+        next: (info) => {
+          this.userInfo = info;
+          this.cdr.detectChanges();
+        }
+    });
 
     this.getAllContacts();
   }
 
   getAllContacts(){
     this.contactService.getAllContacts().subscribe( {
-      next: (contacts => this.userContacts = contacts),
+      next: (contacts => {
+        this.userContacts = contacts;
+        this.cdr.detectChanges();
+      }),
       error: (err: Error) => {
         this.userContacts = [];
         this.error = err.message;
+        this.cdr.detectChanges(); 
       }
     })
   }
@@ -47,7 +51,8 @@ export class ContactPage {
     this.selectedContactId = contactId;
   }
 
-  openContactForm() {
+  openContactForm(action: 'create' | 'update') {
+    this.action = action;
     this.contactFormEnabled = true;
   }
 
@@ -63,14 +68,17 @@ export class ContactPage {
     }
   }
 
-  deleteContact(contactId: number) {
+  deleteContact(contactId: number|undefined) {
+    if (contactId === undefined) return;
     this.contactService.deleteContact(contactId).subscribe({
       next: () => {
         console.log('Contact deleted successfully');
+        this.selectedContactId = undefined;
         this.getAllContacts(); // Refresh the contact list after deletion
       },
       error: (err: Error) => {
         console.error('Error deleting contact:', err.message);
+        this.cdr.detectChanges(); 
       }
     });
   }
